requirements.txt
matplotlib==3.10.8
numpy==2.4.0
pandas==2.3.3
Pillow==12.0.0
torch==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
